Various features (#2)
* feat(module, receivails, transactions, merchant): added custom auth header to secure all controllers

The secret is provided by a secret service, the custom auth header is depending now by a global key.
Also solved problems with the linter and prettier

BREAKING CHANGE: All the controllers includes new auth, the test is securing that feature

* feat(module, receivails, transactions, merchant): added custom auth header to secure all controllers

The secret is provided by a secret service, the custom auth header is depending now by a global key.
Also solved problems with the linter and prettier

BREAKING CHANGE: All the controllers includes new auth, the test is securing that feature

* feat(module, receivails, transactions, merchant): added custom auth header to secure all controllers

The secret is provided by a secret service, the custom auth header is depending now by a global key.
Also solved problems with the linter and prettier

BREAKING CHANGE: All the controllers includes new auth, the test is securing that feature

---------

Co-authored-by: santiago pedrozo <santiago@cashea.app>